requirements.txt
python-dotenv
gradio_client
gradio
git+https://github.com/huggingface/transformers.git
